Updated Datatypes

+++ inc/core/vt_api.h
@@ -4,7 +4,7 @@
 #ifndef _VT_API_H
 #define _VT_API_H
 
-#include "stdio.h"
+#include <stdio.h>
 
 #include "vt_port.h"
 
@@ -12,7 +12,6 @@
 #define VT_ERROR                              0X01
 #define VT_PTR_ERROR                          0X02
 #define VT_MUTEX_ERROR                        0X03
-
 #define VT_UNIDENTIFIED_SENSOR_NAME           0X05
 #define VT_THREAD_SLEEP_ERROR                 0X06
 #define VT_REPEATABILITY_ERROR                0X11
@@ -51,13 +50,13 @@ typedef struct VT_DATABASE_STRUCT
     uint32_t vt_flash_address;
     uint32_t vt_fallcurve_component_id;
 
-    uint32_t _vt_total_fingerprints;
+    uint8_t _vt_total_fingerprints;
     uint32_t _vt_fingerprintdb[10][102];
 
-    uint32_t _vt_total_falltime;
+    uint8_t _vt_total_falltime;
     uint32_t _vt_falltimedb[10][2];
 
-    uint32_t _vt_total_pearson_coefficient;
+    uint8_t _vt_total_pearson_coefficient;
     float _vt_pearson_coefficientdb[10][2];
 
 } VT_DATABASE;
